First commit towards undoable transactions - Mostly refactorings, making access to store records more abstract - Added reference to store and record for Commands - Added RevertibleCommand as the first step towards RevertibleTransaction

+++ kernel/src/main/java/org/neo4j/kernel/impl/nioneo/xa/WriteTransaction.java
@@ -268,7 +268,7 @@ public class WriteTransaction extends XaTransaction implements NeoStoreTransacti
                         }
                     }
                 }
-                removeRelationshipTypeFromCache( record.getId() );
+                removeRelationshipTypeFromCache( (int) record.getId() );
             }
             for ( NodeRecord record : nodeRecords.values() )
             {
@@ -1615,12 +1615,12 @@ public class WriteTransaction extends XaTransaction implements NeoStoreTransacti
 
     void addRelationshipTypeRecord( RelationshipTypeRecord record )
     {
-        relTypeRecords.put( record.getId(), record );
+        relTypeRecords.put( (int) record.getId(), record );
     }
 
     void addPropertyIndexRecord( PropertyIndexRecord record )
     {
-        propIndexRecords.put( record.getId(), record );
+        propIndexRecords.put( (int) record.getId(), record );
     }
 
     PropertyIndexRecord getPropertyIndexRecord( int id )
